Move XMLJournal e XMLIssue de pid_provider para xmlsps e cria XMLSPS em xmlsps. Cria PidProviderXML.current_xml (XMLSPS)
